Restructured and fixed 'nix'
