+++ plutus-core/plutus-core.cabal
@@ -121,8 +121,8 @@ library
 
         Language.UntypedPlutusCore
         Language.UntypedPlutusCore.DeBruijn
+        Language.UntypedPlutusCore.Evaluation.HOAS
         Language.UntypedPlutusCore.Evaluation.Machine.Cek
-        Language.UntypedPlutusCore.HOAS
         Language.UntypedPlutusCore.Parser
         Language.UntypedPlutusCore.Rename
